Move some parsing functions into configuration_parsing.cpp
